DEPRECATE - refactor FE logic/tests into deprecated files (no logic change)
