+++ service/frontend/admin_handler.go
@@ -43,9 +43,7 @@ import (
 	"go.temporal.io/server/common/util"
 
 	"github.com/pborman/uuid"
-	commonpb "go.temporal.io/api/common/v1"
 	enumspb "go.temporal.io/api/enums/v1"
-	historypb "go.temporal.io/api/history/v1"
 	namespacepb "go.temporal.io/api/namespace/v1"
 	"go.temporal.io/api/serviceerror"
 	workflowpb "go.temporal.io/api/workflow/v1"
@@ -58,11 +56,9 @@ import (
 	"go.temporal.io/server/api/adminservice/v1"
 	clusterspb "go.temporal.io/server/api/cluster/v1"
 	enumsspb "go.temporal.io/server/api/enums/v1"
-	historyspb "go.temporal.io/server/api/history/v1"
 	"go.temporal.io/server/api/historyservice/v1"
 	persistencespb "go.temporal.io/server/api/persistence/v1"
 	replicationspb "go.temporal.io/server/api/replication/v1"
-	tokenspb "go.temporal.io/server/api/token/v1"
 	serverClient "go.temporal.io/server/client"
 	"go.temporal.io/server/client/admin"
 	"go.temporal.io/server/client/frontend"
@@ -80,11 +76,9 @@ import (
 	"go.temporal.io/server/common/namespace"
 	"go.temporal.io/server/common/persistence"
 	"go.temporal.io/server/common/persistence/serialization"
-	"go.temporal.io/server/common/persistence/versionhistory"
 	"go.temporal.io/server/common/persistence/visibility/manager"
 	"go.temporal.io/server/common/persistence/visibility/store/elasticsearch"
 	esclient "go.temporal.io/server/common/persistence/visibility/store/elasticsearch/client"
-	"go.temporal.io/server/common/persistence/visibility/store/standard/cassandra"
 	"go.temporal.io/server/common/primitives/timestamp"
 	"go.temporal.io/server/common/sdk"
 	"go.temporal.io/server/common/searchattribute"
@@ -105,30 +99,32 @@ type (
 	AdminHandler struct {
 		status int32
 
-		logger                      log.Logger
-		numberOfHistoryShards       int32
-		ESClient                    esclient.Client
-		config                      *Config
-		namespaceDLQHandler         namespace.DLQMessageHandler
-		eventSerializer             serialization.Serializer
-		visibilityMgr               manager.VisibilityManager
+		logger                     log.Logger
+		numberOfHistoryShards      int32
+		ESClient                   esclient.Client
+		config                     *Config
+		namespaceDLQHandler        namespace.DLQMessageHandler
+		eventSerializer            serialization.Serializer
+		visibilityMgr              manager.VisibilityManager
+		namespaceReplicationQueue  persistence.NamespaceReplicationQueue
+		taskManager                persistence.TaskManager
+		clusterMetadataManager     persistence.ClusterMetadataManager
+		persistenceMetadataManager persistence.MetadataManager
+		clientFactory              serverClient.Factory
+		clientBean                 serverClient.Bean
+		historyClient              historyservice.HistoryServiceClient
+		sdkClientFactory           sdk.ClientFactory
+		membershipMonitor          membership.Monitor
+		hostInfoProvider           membership.HostInfoProvider
+		metricsHandler             metrics.Handler
+		namespaceRegistry          namespace.Registry
+		saProvider                 searchattribute.Provider
+		saManager                  searchattribute.Manager
+		clusterMetadata            cluster.Metadata
+		healthServer               *health.Server
+
+		// DEPRECATED
 		persistenceExecutionManager persistence.ExecutionManager
-		namespaceReplicationQueue   persistence.NamespaceReplicationQueue
-		taskManager                 persistence.TaskManager
-		clusterMetadataManager      persistence.ClusterMetadataManager
-		persistenceMetadataManager  persistence.MetadataManager
-		clientFactory               serverClient.Factory
-		clientBean                  serverClient.Bean
-		historyClient               historyservice.HistoryServiceClient
-		sdkClientFactory            sdk.ClientFactory
-		membershipMonitor           membership.Monitor
-		hostInfoProvider            membership.HostInfoProvider
-		metricsHandler              metrics.Handler
-		namespaceRegistry           namespace.Registry
-		saProvider                  searchattribute.Provider
-		saManager                   searchattribute.Manager
-		clusterMetadata             cluster.Metadata
-		healthServer                *health.Server
 	}
 
 	NewAdminHandlerArgs struct {
@@ -139,7 +135,6 @@ type (
 		EsClient                            esclient.Client
 		VisibilityMrg                       manager.VisibilityManager
 		Logger                              log.Logger
-		PersistenceExecutionManager         persistence.ExecutionManager
 		TaskManager                         persistence.TaskManager
 		ClusterMetadataManager              persistence.ClusterMetadataManager
 		PersistenceMetadataManager          persistence.MetadataManager
@@ -159,6 +154,9 @@ type (
 		HealthServer                        *health.Server
 		EventSerializer                     serialization.Serializer
 		TimeSource                          clock.TimeSource
+
+		// DEPRECATED
+		PersistenceExecutionManager persistence.ExecutionManager
 	}
 )
 
@@ -897,117 +895,7 @@ func (adh *AdminHandler) GetWorkflowExecutionRawHistoryV2(ctx context.Context, r
 		return nil, err
 	}
 
-	ns, err := adh.namespaceRegistry.GetNamespaceByID(namespace.ID(request.GetNamespaceId()))
-	if err != nil {
-		return nil, err
-	}
-
-	execution := request.Execution
-	var pageToken *tokenspb.RawHistoryContinuation
-	var targetVersionHistory *historyspb.VersionHistory
-	if request.NextPageToken == nil {
-		response, err := adh.historyClient.GetMutableState(ctx, &historyservice.GetMutableStateRequest{
-			NamespaceId: ns.ID().String(),
-			Execution:   execution,
-		})
-		if err != nil {
-			return nil, err
-		}
-
-		targetVersionHistory, err = adh.setRequestDefaultValueAndGetTargetVersionHistory(
-			request,
-			response.GetVersionHistories(),
-		)
-		if err != nil {
-			return nil, err
-		}
-
-		pageToken = generatePaginationToken(request, response.GetVersionHistories())
-	} else {
-		pageToken, err = deserializeRawHistoryToken(request.NextPageToken)
-		if err != nil {
-			return nil, err
-		}
-		versionHistories := pageToken.GetVersionHistories()
-		if versionHistories == nil {
-			return nil, errInvalidVersionHistories
-		}
-		targetVersionHistory, err = adh.setRequestDefaultValueAndGetTargetVersionHistory(
-			request,
-			versionHistories,
-		)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	if err := validatePaginationToken(
-		request,
-		pageToken,
-	); err != nil {
-		return nil, err
-	}
-
-	if pageToken.GetStartEventId()+1 == pageToken.GetEndEventId() {
-		// API is exclusive-exclusive. Return empty response here.
-		return &adminservice.GetWorkflowExecutionRawHistoryV2Response{
-			HistoryBatches: []*commonpb.DataBlob{},
-			NextPageToken:  nil, // no further pagination
-			VersionHistory: targetVersionHistory,
-		}, nil
-	}
-	pageSize := int(request.GetMaximumPageSize())
-	shardID := common.WorkflowIDToHistoryShard(
-		ns.ID().String(),
-		execution.GetWorkflowId(),
-		adh.numberOfHistoryShards,
-	)
-	rawHistoryResponse, err := adh.persistenceExecutionManager.ReadRawHistoryBranch(ctx, &persistence.ReadHistoryBranchRequest{
-		BranchToken: targetVersionHistory.GetBranchToken(),
-		// GetWorkflowExecutionRawHistoryV2 is exclusive exclusive.
-		// ReadRawHistoryBranch is inclusive exclusive.
-		MinEventID:    pageToken.GetStartEventId() + 1,
-		MaxEventID:    pageToken.GetEndEventId(),
-		PageSize:      pageSize,
-		NextPageToken: pageToken.PersistenceToken,
-		ShardID:       shardID,
-	})
-	if err != nil {
-		if _, isNotFound := err.(*serviceerror.NotFound); isNotFound {
-			// when no events can be returned from DB, DB layer will return
-			// EntityNotExistsError, this API shall return empty response
-			return &adminservice.GetWorkflowExecutionRawHistoryV2Response{
-				HistoryBatches: []*commonpb.DataBlob{},
-				NextPageToken:  nil, // no further pagination
-				VersionHistory: targetVersionHistory,
-			}, nil
-		}
-		return nil, err
-	}
-
-	pageToken.PersistenceToken = rawHistoryResponse.NextPageToken
-	size := rawHistoryResponse.Size
-	adh.metricsHandler.Histogram(metrics.HistorySize.GetMetricName(), metrics.HistorySize.GetMetricUnit()).Record(
-		int64(size),
-		metrics.NamespaceTag(ns.Name().String()),
-		metrics.OperationTag(metrics.AdminGetWorkflowExecutionRawHistoryV2Scope),
-	)
-
-	result := &adminservice.GetWorkflowExecutionRawHistoryV2Response{
-		HistoryBatches: rawHistoryResponse.HistoryEventBlobs,
-		VersionHistory: targetVersionHistory,
-		HistoryNodeIds: rawHistoryResponse.NodeIDs,
-	}
-	if len(pageToken.PersistenceToken) == 0 {
-		result.NextPageToken = nil
-	} else {
-		result.NextPageToken, err = serializeRawHistoryToken(pageToken)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return result, nil
+	return adh.getWorkflowExecutionRawHistoryV2(ctx, request)
 }
 
 // DescribeCluster return information about a temporal cluster
@@ -1642,160 +1530,7 @@ func (adh *AdminHandler) DeleteWorkflowExecution(
 		return nil, err
 	}
 
-	namespaceID, err := adh.namespaceRegistry.GetNamespaceID(namespace.Name(request.GetNamespace()))
-	if err != nil {
-		return nil, err
-	}
-	execution := request.Execution
-
-	shardID := common.WorkflowIDToHistoryShard(
-		namespaceID.String(),
-		execution.GetWorkflowId(),
-		adh.numberOfHistoryShards,
-	)
-	logger := log.With(adh.logger,
-		tag.WorkflowNamespace(request.Namespace),
-		tag.WorkflowID(execution.WorkflowId),
-		tag.WorkflowRunID(execution.RunId),
-	)
-
-	if execution.RunId == "" {
-		resp, err := adh.persistenceExecutionManager.GetCurrentExecution(ctx, &persistence.GetCurrentExecutionRequest{
-			ShardID:     shardID,
-			NamespaceID: namespaceID.String(),
-			WorkflowID:  execution.WorkflowId,
-		})
-		if err != nil {
-			return nil, err
-		}
-		execution.RunId = resp.RunID
-	}
-
-	var warnings []string
-	var branchTokens [][]byte
-	var startTime, closeTime *time.Time
-	cassVisBackend := adh.visibilityMgr.HasStoreName(cassandra.CassandraPersistenceName)
-
-	resp, err := adh.persistenceExecutionManager.GetWorkflowExecution(ctx, &persistence.GetWorkflowExecutionRequest{
-		ShardID:     shardID,
-		NamespaceID: namespaceID.String(),
-		WorkflowID:  execution.WorkflowId,
-		RunID:       execution.RunId,
-	})
-	if err != nil {
-		if common.IsContextCanceledErr(err) || common.IsContextDeadlineExceededErr(err) {
-			return nil, err
-		}
-		// continue to deletion
-		warnMsg := "Unable to load mutable state when deleting workflow execution, " +
-			"will skip deleting workflow history and cassandra visibility record"
-		logger.Warn(warnMsg, tag.Error(err))
-		warnings = append(warnings, fmt.Sprintf("%s. Error: %v", warnMsg, err.Error()))
-	} else {
-		// load necessary information from mutable state
-		executionInfo := resp.State.GetExecutionInfo()
-		histories := executionInfo.GetVersionHistories().GetHistories()
-		branchTokens = make([][]byte, 0, len(histories))
-		for _, historyItem := range histories {
-			branchTokens = append(branchTokens, historyItem.GetBranchToken())
-		}
-
-		if cassVisBackend {
-			if resp.State.ExecutionState.State != enumsspb.WORKFLOW_EXECUTION_STATE_COMPLETED {
-				startTime = executionInfo.GetStartTime()
-			} else if executionInfo.GetCloseTime() != nil {
-				closeTime = executionInfo.GetCloseTime()
-			} else {
-				completionEvent, err := adh.getWorkflowCompletionEvent(ctx, shardID, resp.State)
-				if err != nil {
-					warnMsg := "Unable to load workflow completion event, will skip deleting visibility record"
-					adh.logger.Warn(warnMsg, tag.Error(err))
-					warnings = append(warnings, fmt.Sprintf("%s. Error: %v", warnMsg, err.Error()))
-				} else {
-					closeTime = completionEvent.GetEventTime()
-				}
-			}
-		}
-	}
-
-	if !cassVisBackend || (startTime != nil || closeTime != nil) {
-		// if using cass visibility, then either start or close time should be non-nil
-		// NOTE: the deletion is best effort, for sql and cassandra visibility implementation,
-		// we can't guarantee there's no update or record close request for this workflow since
-		// visibility queue processing is async. Operator can call this api again to delete visibility
-		// record again if this happens.
-		if _, err := adh.historyClient.DeleteWorkflowVisibilityRecord(ctx, &historyservice.DeleteWorkflowVisibilityRecordRequest{
-			NamespaceId:       namespaceID.String(),
-			Execution:         execution,
-			WorkflowStartTime: startTime,
-			WorkflowCloseTime: closeTime,
-		}); err != nil {
-			return nil, err
-		}
-	}
-
-	if err := adh.persistenceExecutionManager.DeleteCurrentWorkflowExecution(ctx, &persistence.DeleteCurrentWorkflowExecutionRequest{
-		ShardID:     shardID,
-		NamespaceID: namespaceID.String(),
-		WorkflowID:  execution.WorkflowId,
-		RunID:       execution.RunId,
-	}); err != nil {
-		return nil, err
-	}
-
-	if err := adh.persistenceExecutionManager.DeleteWorkflowExecution(ctx, &persistence.DeleteWorkflowExecutionRequest{
-		ShardID:     shardID,
-		NamespaceID: namespaceID.String(),
-		WorkflowID:  execution.WorkflowId,
-		RunID:       execution.RunId,
-	}); err != nil {
-		return nil, err
-	}
-
-	for _, branchToken := range branchTokens {
-		if err := adh.persistenceExecutionManager.DeleteHistoryBranch(ctx, &persistence.DeleteHistoryBranchRequest{
-			ShardID:     shardID,
-			BranchToken: branchToken,
-		}); err != nil {
-			warnMsg := "Failed to delete history branch, skip"
-			adh.logger.Warn(warnMsg, tag.WorkflowBranchID(string(branchToken)), tag.Error(err))
-			warnings = append(warnings, fmt.Sprintf("%s. BranchToken: %v, Error: %v", warnMsg, branchToken, err.Error()))
-		}
-	}
-
-	return &adminservice.DeleteWorkflowExecutionResponse{
-		Warnings: warnings,
-	}, nil
-}
-
-func (adh *AdminHandler) validateGetWorkflowExecutionRawHistoryV2Request(
-	request *adminservice.GetWorkflowExecutionRawHistoryV2Request,
-) error {
-
-	execution := request.Execution
-	if execution.GetWorkflowId() == "" {
-		return errWorkflowIDNotSet
-	}
-	// TODO currently, this API is only going to be used by re-send history events
-	// to remote cluster if kafka is lossy again, in the future, this API can be used
-	// by CLI and client, then empty runID (meaning the current workflow) should be allowed
-	if execution.GetRunId() == "" || uuid.Parse(execution.GetRunId()) == nil {
-		return errInvalidRunID
-	}
-
-	pageSize := int(request.GetMaximumPageSize())
-	if pageSize <= 0 {
-		return errInvalidPageSize
-	}
-
-	if request.GetStartEventId() == common.EmptyEventID &&
-		request.GetStartEventVersion() == common.EmptyVersion &&
-		request.GetEndEventId() == common.EmptyEventID &&
-		request.GetEndEventVersion() == common.EmptyVersion {
-		return errInvalidEventQueryRange
-	}
-
-	return nil
+	return adh.deleteWorkflowExecution(ctx, request)
 }
 
 func (adh *AdminHandler) validateRemoteClusterMetadata(metadata *adminservice.DescribeClusterResponse) error {
@@ -1832,124 +1567,6 @@ func (adh *AdminHandler) validateRemoteClusterMetadata(metadata *adminservice.De
 		}
 	}
 	return nil
-}
-
-func (adh *AdminHandler) setRequestDefaultValueAndGetTargetVersionHistory(
-	request *adminservice.GetWorkflowExecutionRawHistoryV2Request,
-	versionHistories *historyspb.VersionHistories,
-) (*historyspb.VersionHistory, error) {
-
-	targetBranch, err := versionhistory.GetCurrentVersionHistory(versionHistories)
-	if err != nil {
-		return nil, err
-	}
-	firstItem, err := versionhistory.GetFirstVersionHistoryItem(targetBranch)
-	if err != nil {
-		return nil, err
-	}
-	lastItem, err := versionhistory.GetLastVersionHistoryItem(targetBranch)
-	if err != nil {
-		return nil, err
-	}
-
-	if request.GetStartEventId() == common.EmptyVersion || request.GetStartEventVersion() == common.EmptyVersion {
-		// If start event is not set, get the events from the first event
-		// As the API is exclusive-exclusive, use first event id - 1 here
-		request.StartEventId = common.FirstEventID - 1
-		request.StartEventVersion = firstItem.GetVersion()
-	}
-	if request.GetEndEventId() == common.EmptyEventID || request.GetEndEventVersion() == common.EmptyVersion {
-		// If end event is not set, get the events until the end event
-		// As the API is exclusive-exclusive, use end event id + 1 here
-		request.EndEventId = lastItem.GetEventId() + 1
-		request.EndEventVersion = lastItem.GetVersion()
-	}
-
-	if request.GetStartEventId() < 0 {
-		return nil, errInvalidFirstNextEventCombination
-	}
-
-	// get branch based on the end event if end event is defined in the request
-	if request.GetEndEventId() == lastItem.GetEventId()+1 &&
-		request.GetEndEventVersion() == lastItem.GetVersion() {
-		// this is a special case, target branch remains the same
-	} else {
-		endItem := versionhistory.NewVersionHistoryItem(request.GetEndEventId(), request.GetEndEventVersion())
-		idx, err := versionhistory.FindFirstVersionHistoryIndexByVersionHistoryItem(versionHistories, endItem)
-		if err != nil {
-			return nil, err
-		}
-
-		targetBranch, err = versionhistory.GetVersionHistory(versionHistories, idx)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	startItem := versionhistory.NewVersionHistoryItem(request.GetStartEventId(), request.GetStartEventVersion())
-	// If the request start event is defined. The start event may be on a different branch as current branch.
-	// We need to find the LCA of the start event and the current branch.
-	if request.GetStartEventId() == common.FirstEventID-1 &&
-		request.GetStartEventVersion() == firstItem.GetVersion() {
-		// this is a special case, start event is on the same branch as target branch
-	} else {
-		if !versionhistory.ContainsVersionHistoryItem(targetBranch, startItem) {
-			idx, err := versionhistory.FindFirstVersionHistoryIndexByVersionHistoryItem(versionHistories, startItem)
-			if err != nil {
-				return nil, err
-			}
-			startBranch, err := versionhistory.GetVersionHistory(versionHistories, idx)
-			if err != nil {
-				return nil, err
-			}
-			startItem, err = versionhistory.FindLCAVersionHistoryItem(targetBranch, startBranch)
-			if err != nil {
-				return nil, err
-			}
-			request.StartEventId = startItem.GetEventId()
-			request.StartEventVersion = startItem.GetVersion()
-		}
-	}
-
-	return targetBranch, nil
-}
-
-func (adh *AdminHandler) getWorkflowCompletionEvent(
-	ctx context.Context,
-	shardID int32,
-	mutableState *persistencespb.WorkflowMutableState,
-) (*historypb.HistoryEvent, error) {
-	executionInfo := mutableState.GetExecutionInfo()
-	completionEventID := mutableState.GetNextEventId() - 1
-
-	currentVersionHistory, err := versionhistory.GetCurrentVersionHistory(executionInfo.VersionHistories)
-	if err != nil {
-		return nil, err
-	}
-	version, err := versionhistory.GetVersionHistoryEventVersion(currentVersionHistory, completionEventID)
-	if err != nil {
-		return nil, err
-	}
-
-	resp, err := adh.persistenceExecutionManager.ReadHistoryBranch(ctx, &persistence.ReadHistoryBranchRequest{
-		ShardID:     shardID,
-		BranchToken: currentVersionHistory.GetBranchToken(),
-		MinEventID:  executionInfo.CompletionEventBatchId,
-		MaxEventID:  completionEventID + 1,
-		PageSize:    1,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	// find history event from batch and return back single event to caller
-	for _, e := range resp.HistoryEvents {
-		if e.EventId == completionEventID && e.Version == version {
-			return e, nil
-		}
-	}
-
-	return nil, serviceerror.NewInternal("Unable to find closed event for workflow")
 }
 
 func (adh *AdminHandler) StreamWorkflowReplicationMessages(
