+++ service/frontend/admin_handler_test.go
@@ -31,48 +31,40 @@ import (
 	"math/rand"
 	"sync"
 	"testing"
-	"time"
 
 	namespacepb "go.temporal.io/api/namespace/v1"
 	"google.golang.org/grpc/metadata"
 
 	historyclient "go.temporal.io/server/client/history"
 	"go.temporal.io/server/common/clock"
-	"go.temporal.io/server/common/persistence/visibility/store/standard/cassandra"
 	"go.temporal.io/server/common/primitives"
+	"go.temporal.io/server/common/primitives/timestamp"
 	"go.temporal.io/server/common/resourcetest"
 
 	"google.golang.org/grpc/health"
 
 	"go.temporal.io/server/api/adminservicemock/v1"
-	"go.temporal.io/server/api/enums/v1"
 	persistencespb "go.temporal.io/server/api/persistence/v1"
 	"go.temporal.io/server/common/cluster"
 	"go.temporal.io/server/common/membership"
-	"go.temporal.io/server/common/primitives/timestamp"
 	"go.temporal.io/server/common/testing/mocksdk"
 
 	"github.com/golang/mock/gomock"
 	"github.com/pborman/uuid"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
-	commonpb "go.temporal.io/api/common/v1"
 	enumspb "go.temporal.io/api/enums/v1"
-	historypb "go.temporal.io/api/history/v1"
 	"go.temporal.io/api/serviceerror"
 	"go.temporal.io/api/workflowservice/v1"
 
 	"go.temporal.io/server/api/adminservice/v1"
-	historyspb "go.temporal.io/server/api/history/v1"
 	"go.temporal.io/server/api/historyservice/v1"
 	"go.temporal.io/server/api/historyservicemock/v1"
 	clientmocks "go.temporal.io/server/client"
-	"go.temporal.io/server/common"
 	"go.temporal.io/server/common/config"
 	"go.temporal.io/server/common/namespace"
 	"go.temporal.io/server/common/persistence"
 	"go.temporal.io/server/common/persistence/serialization"
-	"go.temporal.io/server/common/persistence/versionhistory"
 	"go.temporal.io/server/common/persistence/visibility/manager"
 	"go.temporal.io/server/common/persistence/visibility/store/elasticsearch"
 	"go.temporal.io/server/common/searchattribute"
@@ -88,6 +80,7 @@ type (
 		mockHistoryClient  *historyservicemock.MockHistoryServiceClient
 		mockNamespaceCache *namespace.MockRegistry
 
+		// DEPRECATED
 		mockExecutionMgr           *persistence.MockExecutionManager
 		mockVisibilityMgr          *manager.MockVisibilityManager
 		mockClusterMetadataManager *persistence.MockClusterMetadataManager
@@ -152,7 +145,6 @@ func (s *adminHandlerSuite) SetupTest() {
 		s.mockResource.ESClient,
 		s.mockResource.GetVisibilityManager(),
 		s.mockResource.GetLogger(),
-		s.mockResource.GetExecutionManager(),
 		s.mockResource.GetTaskManager(),
 		s.mockResource.GetClusterMetadataManager(),
 		s.mockResource.GetMetadataManager(),
@@ -172,6 +164,7 @@ func (s *adminHandlerSuite) SetupTest() {
 		health.NewServer(),
 		serialization.NewSerializer(),
 		clock.NewRealTimeSource(),
+		s.mockResource.GetExecutionManager(),
 	}
 	s.mockMetadata.EXPECT().GetCurrentClusterName().Return(uuid.New()).AnyTimes()
 	s.handler = NewAdminHandler(args)
@@ -181,290 +174,6 @@ func (s *adminHandlerSuite) SetupTest() {
 func (s *adminHandlerSuite) TearDownTest() {
 	s.controller.Finish()
 	s.handler.Stop()
-}
-
-func (s *adminHandlerSuite) Test_GetWorkflowExecutionRawHistoryV2_FailedOnInvalidWorkflowID() {
-	ctx := context.Background()
-	_, err := s.handler.GetWorkflowExecutionRawHistoryV2(ctx,
-		&adminservice.GetWorkflowExecutionRawHistoryV2Request{
-			NamespaceId: s.namespaceID.String(),
-			Execution: &commonpb.WorkflowExecution{
-				WorkflowId: "",
-				RunId:      uuid.New(),
-			},
-			StartEventId:      1,
-			StartEventVersion: 100,
-			EndEventId:        10,
-			EndEventVersion:   100,
-			MaximumPageSize:   1,
-			NextPageToken:     nil,
-		})
-	s.Error(err)
-}
-
-func (s *adminHandlerSuite) Test_GetWorkflowExecutionRawHistoryV2_FailedOnInvalidRunID() {
-	ctx := context.Background()
-	_, err := s.handler.GetWorkflowExecutionRawHistoryV2(ctx,
-		&adminservice.GetWorkflowExecutionRawHistoryV2Request{
-			NamespaceId: s.namespaceID.String(),
-			Execution: &commonpb.WorkflowExecution{
-				WorkflowId: "workflowID",
-				RunId:      "runID",
-			},
-			StartEventId:      1,
-			StartEventVersion: 100,
-			EndEventId:        10,
-			EndEventVersion:   100,
-			MaximumPageSize:   1,
-			NextPageToken:     nil,
-		})
-	s.Error(err)
-}
-
-func (s *adminHandlerSuite) Test_GetWorkflowExecutionRawHistoryV2_FailedOnInvalidSize() {
-	ctx := context.Background()
-	_, err := s.handler.GetWorkflowExecutionRawHistoryV2(ctx,
-		&adminservice.GetWorkflowExecutionRawHistoryV2Request{
-			NamespaceId: s.namespaceID.String(),
-			Execution: &commonpb.WorkflowExecution{
-				WorkflowId: "workflowID",
-				RunId:      uuid.New(),
-			},
-			StartEventId:      1,
-			StartEventVersion: 100,
-			EndEventId:        10,
-			EndEventVersion:   100,
-			MaximumPageSize:   -1,
-			NextPageToken:     nil,
-		})
-	s.Error(err)
-}
-
-func (s *adminHandlerSuite) Test_GetWorkflowExecutionRawHistoryV2_FailedOnNamespaceCache() {
-	ctx := context.Background()
-	s.mockNamespaceCache.EXPECT().GetNamespaceByID(s.namespaceID).Return(nil, fmt.Errorf("test"))
-	_, err := s.handler.GetWorkflowExecutionRawHistoryV2(ctx,
-		&adminservice.GetWorkflowExecutionRawHistoryV2Request{
-			NamespaceId: s.namespaceID.String(),
-			Execution: &commonpb.WorkflowExecution{
-				WorkflowId: "workflowID",
-				RunId:      uuid.New(),
-			},
-			StartEventId:      1,
-			StartEventVersion: 100,
-			EndEventId:        10,
-			EndEventVersion:   100,
-			MaximumPageSize:   1,
-			NextPageToken:     nil,
-		})
-	s.Error(err)
-}
-
-func (s *adminHandlerSuite) Test_GetWorkflowExecutionRawHistoryV2() {
-	ctx := context.Background()
-	s.mockNamespaceCache.EXPECT().GetNamespaceByID(s.namespaceID).Return(s.namespaceEntry, nil).AnyTimes()
-	branchToken := []byte{1}
-	versionHistory := versionhistory.NewVersionHistory(branchToken, []*historyspb.VersionHistoryItem{
-		versionhistory.NewVersionHistoryItem(int64(10), int64(100)),
-	})
-	versionHistories := versionhistory.NewVersionHistories(versionHistory)
-	mState := &historyservice.GetMutableStateResponse{
-		NextEventId:        11,
-		CurrentBranchToken: branchToken,
-		VersionHistories:   versionHistories,
-	}
-	s.mockHistoryClient.EXPECT().GetMutableState(gomock.Any(), gomock.Any()).Return(mState, nil).AnyTimes()
-
-	s.mockExecutionMgr.EXPECT().ReadRawHistoryBranch(gomock.Any(), gomock.Any()).Return(&persistence.ReadRawHistoryBranchResponse{
-		HistoryEventBlobs: []*commonpb.DataBlob{},
-		NextPageToken:     []byte{},
-		Size:              0,
-	}, nil)
-	_, err := s.handler.GetWorkflowExecutionRawHistoryV2(ctx,
-		&adminservice.GetWorkflowExecutionRawHistoryV2Request{
-			NamespaceId: s.namespaceID.String(),
-			Execution: &commonpb.WorkflowExecution{
-				WorkflowId: "workflowID",
-				RunId:      uuid.New(),
-			},
-			StartEventId:      1,
-			StartEventVersion: 100,
-			EndEventId:        10,
-			EndEventVersion:   100,
-			MaximumPageSize:   10,
-			NextPageToken:     nil,
-		})
-	s.NoError(err)
-}
-
-func (s *adminHandlerSuite) Test_GetWorkflowExecutionRawHistoryV2_SameStartIDAndEndID() {
-	ctx := context.Background()
-	s.mockNamespaceCache.EXPECT().GetNamespaceByID(s.namespaceID).Return(s.namespaceEntry, nil).AnyTimes()
-	branchToken := []byte{1}
-	versionHistory := versionhistory.NewVersionHistory(branchToken, []*historyspb.VersionHistoryItem{
-		versionhistory.NewVersionHistoryItem(int64(10), int64(100)),
-	})
-	versionHistories := versionhistory.NewVersionHistories(versionHistory)
-	mState := &historyservice.GetMutableStateResponse{
-		NextEventId:        11,
-		CurrentBranchToken: branchToken,
-		VersionHistories:   versionHistories,
-	}
-	s.mockHistoryClient.EXPECT().GetMutableState(gomock.Any(), gomock.Any()).Return(mState, nil).AnyTimes()
-
-	resp, err := s.handler.GetWorkflowExecutionRawHistoryV2(ctx,
-		&adminservice.GetWorkflowExecutionRawHistoryV2Request{
-			NamespaceId: s.namespaceID.String(),
-			Execution: &commonpb.WorkflowExecution{
-				WorkflowId: "workflowID",
-				RunId:      uuid.New(),
-			},
-			StartEventId:      10,
-			StartEventVersion: 100,
-			EndEventId:        common.EmptyEventID,
-			EndEventVersion:   common.EmptyVersion,
-			MaximumPageSize:   1,
-			NextPageToken:     nil,
-		})
-	s.Nil(resp.NextPageToken)
-	s.NoError(err)
-}
-
-func (s *adminHandlerSuite) Test_SetRequestDefaultValueAndGetTargetVersionHistory_DefinedStartAndEnd() {
-	inputStartEventID := int64(1)
-	inputStartVersion := int64(10)
-	inputEndEventID := int64(100)
-	inputEndVersion := int64(11)
-	firstItem := versionhistory.NewVersionHistoryItem(inputStartEventID, inputStartVersion)
-	endItem := versionhistory.NewVersionHistoryItem(inputEndEventID, inputEndVersion)
-	versionHistory := versionhistory.NewVersionHistory([]byte{}, []*historyspb.VersionHistoryItem{firstItem, endItem})
-	versionHistories := versionhistory.NewVersionHistories(versionHistory)
-	request := &adminservice.GetWorkflowExecutionRawHistoryV2Request{
-		NamespaceId: s.namespaceID.String(),
-		Execution: &commonpb.WorkflowExecution{
-			WorkflowId: "workflowID",
-			RunId:      uuid.New(),
-		},
-		StartEventId:      inputStartEventID,
-		StartEventVersion: inputStartVersion,
-		EndEventId:        inputEndEventID,
-		EndEventVersion:   inputEndVersion,
-		MaximumPageSize:   10,
-		NextPageToken:     nil,
-	}
-
-	targetVersionHistory, err := s.handler.setRequestDefaultValueAndGetTargetVersionHistory(
-		request,
-		versionHistories,
-	)
-	s.Equal(request.GetStartEventId(), inputStartEventID)
-	s.Equal(request.GetEndEventId(), inputEndEventID)
-	s.Equal(targetVersionHistory, versionHistory)
-	s.NoError(err)
-}
-
-func (s *adminHandlerSuite) Test_SetRequestDefaultValueAndGetTargetVersionHistory_DefinedEndEvent() {
-	inputStartEventID := int64(1)
-	inputEndEventID := int64(100)
-	inputStartVersion := int64(10)
-	inputEndVersion := int64(11)
-	firstItem := versionhistory.NewVersionHistoryItem(inputStartEventID, inputStartVersion)
-	targetItem := versionhistory.NewVersionHistoryItem(inputEndEventID, inputEndVersion)
-	versionHistory := versionhistory.NewVersionHistory([]byte{}, []*historyspb.VersionHistoryItem{firstItem, targetItem})
-	versionHistories := versionhistory.NewVersionHistories(versionHistory)
-	request := &adminservice.GetWorkflowExecutionRawHistoryV2Request{
-		NamespaceId: s.namespaceID.String(),
-		Execution: &commonpb.WorkflowExecution{
-			WorkflowId: "workflowID",
-			RunId:      uuid.New(),
-		},
-		StartEventId:      common.EmptyEventID,
-		StartEventVersion: common.EmptyVersion,
-		EndEventId:        inputEndEventID,
-		EndEventVersion:   inputEndVersion,
-		MaximumPageSize:   10,
-		NextPageToken:     nil,
-	}
-
-	targetVersionHistory, err := s.handler.setRequestDefaultValueAndGetTargetVersionHistory(
-		request,
-		versionHistories,
-	)
-	s.Equal(request.GetStartEventId(), inputStartEventID-1)
-	s.Equal(request.GetEndEventId(), inputEndEventID)
-	s.Equal(targetVersionHistory, versionHistory)
-	s.NoError(err)
-}
-
-func (s *adminHandlerSuite) Test_SetRequestDefaultValueAndGetTargetVersionHistory_DefinedStartEvent() {
-	inputStartEventID := int64(1)
-	inputEndEventID := int64(100)
-	inputStartVersion := int64(10)
-	inputEndVersion := int64(11)
-	firstItem := versionhistory.NewVersionHistoryItem(inputStartEventID, inputStartVersion)
-	targetItem := versionhistory.NewVersionHistoryItem(inputEndEventID, inputEndVersion)
-	versionHistory := versionhistory.NewVersionHistory([]byte{}, []*historyspb.VersionHistoryItem{firstItem, targetItem})
-	versionHistories := versionhistory.NewVersionHistories(versionHistory)
-	request := &adminservice.GetWorkflowExecutionRawHistoryV2Request{
-		NamespaceId: s.namespaceID.String(),
-		Execution: &commonpb.WorkflowExecution{
-			WorkflowId: "workflowID",
-			RunId:      uuid.New(),
-		},
-		StartEventId:      inputStartEventID,
-		StartEventVersion: inputStartVersion,
-		EndEventId:        common.EmptyEventID,
-		EndEventVersion:   common.EmptyVersion,
-		MaximumPageSize:   10,
-		NextPageToken:     nil,
-	}
-
-	targetVersionHistory, err := s.handler.setRequestDefaultValueAndGetTargetVersionHistory(
-		request,
-		versionHistories,
-	)
-	s.Equal(request.GetStartEventId(), inputStartEventID)
-	s.Equal(request.GetEndEventId(), inputEndEventID+1)
-	s.Equal(targetVersionHistory, versionHistory)
-	s.NoError(err)
-}
-
-func (s *adminHandlerSuite) Test_SetRequestDefaultValueAndGetTargetVersionHistory_NonCurrentBranch() {
-	inputStartEventID := int64(1)
-	inputEndEventID := int64(100)
-	inputStartVersion := int64(10)
-	inputEndVersion := int64(101)
-	item1 := versionhistory.NewVersionHistoryItem(inputStartEventID, inputStartVersion)
-	item2 := versionhistory.NewVersionHistoryItem(inputEndEventID, inputEndVersion)
-	versionHistory1 := versionhistory.NewVersionHistory([]byte{}, []*historyspb.VersionHistoryItem{item1, item2})
-	item3 := versionhistory.NewVersionHistoryItem(int64(10), int64(20))
-	item4 := versionhistory.NewVersionHistoryItem(int64(20), int64(51))
-	versionHistory2 := versionhistory.NewVersionHistory([]byte{}, []*historyspb.VersionHistoryItem{item1, item3, item4})
-	versionHistories := versionhistory.NewVersionHistories(versionHistory1)
-	_, _, err := versionhistory.AddVersionHistory(versionHistories, versionHistory2)
-	s.NoError(err)
-	request := &adminservice.GetWorkflowExecutionRawHistoryV2Request{
-		NamespaceId: s.namespaceID.String(),
-		Execution: &commonpb.WorkflowExecution{
-			WorkflowId: "workflowID",
-			RunId:      uuid.New(),
-		},
-		StartEventId:      9,
-		StartEventVersion: 20,
-		EndEventId:        inputEndEventID,
-		EndEventVersion:   inputEndVersion,
-		MaximumPageSize:   10,
-		NextPageToken:     nil,
-	}
-
-	targetVersionHistory, err := s.handler.setRequestDefaultValueAndGetTargetVersionHistory(
-		request,
-		versionHistories,
-	)
-	s.Equal(request.GetStartEventId(), inputStartEventID)
-	s.Equal(request.GetEndEventId(), inputEndEventID)
-	s.Equal(targetVersionHistory, versionHistory1)
-	s.NoError(err)
 }
 
 func (s *adminHandlerSuite) Test_AddSearchAttributes() {
@@ -1252,204 +961,6 @@ func (s *adminHandlerSuite) Test_ListClusters_Success() {
 	s.NoError(err)
 	s.Equal(1, len(resp.Clusters))
 	s.Equal(0, len(resp.GetNextPageToken()))
-}
-
-func (s *adminHandlerSuite) TestDeleteWorkflowExecution_DeleteCurrentExecution() {
-	execution := commonpb.WorkflowExecution{
-		WorkflowId: "workflowID",
-	}
-
-	request := &adminservice.DeleteWorkflowExecutionRequest{
-		Namespace: s.namespace.String(),
-		Execution: &execution,
-	}
-
-	s.mockNamespaceCache.EXPECT().GetNamespaceID(s.namespace).Return(s.namespaceID, nil).AnyTimes()
-	s.mockVisibilityMgr.EXPECT().HasStoreName(cassandra.CassandraPersistenceName).Return(false)
-
-	s.mockExecutionMgr.EXPECT().GetCurrentExecution(gomock.Any(), gomock.Any()).Return(nil, errors.New("some random error"))
-	resp, err := s.handler.DeleteWorkflowExecution(context.Background(), request)
-	s.Nil(resp)
-	s.Error(err)
-
-	mutableState := &persistencespb.WorkflowMutableState{
-		ExecutionInfo: &persistencespb.WorkflowExecutionInfo{
-			VersionHistories: &historyspb.VersionHistories{
-				CurrentVersionHistoryIndex: 0,
-				Histories: []*historyspb.VersionHistory{
-					{BranchToken: []byte("branch1")},
-					{BranchToken: []byte("branch2")},
-					{BranchToken: []byte("branch3")},
-				},
-			},
-		},
-	}
-
-	shardID := common.WorkflowIDToHistoryShard(
-		s.namespaceID.String(),
-		execution.GetWorkflowId(),
-		s.handler.numberOfHistoryShards,
-	)
-	runID := uuid.New()
-	s.mockExecutionMgr.EXPECT().GetCurrentExecution(gomock.Any(), gomock.Any()).Return(&persistence.GetCurrentExecutionResponse{
-		StartRequestID: uuid.New(),
-		RunID:          runID,
-		State:          enums.WORKFLOW_EXECUTION_STATE_COMPLETED,
-		Status:         enumspb.WORKFLOW_EXECUTION_STATUS_COMPLETED,
-	}, nil)
-	s.mockExecutionMgr.EXPECT().GetWorkflowExecution(gomock.Any(), &persistence.GetWorkflowExecutionRequest{
-		ShardID:     shardID,
-		NamespaceID: s.namespaceID.String(),
-		WorkflowID:  execution.WorkflowId,
-		RunID:       runID,
-	}).Return(&persistence.GetWorkflowExecutionResponse{State: mutableState}, nil)
-	s.mockHistoryClient.EXPECT().DeleteWorkflowVisibilityRecord(gomock.Any(), &historyservice.DeleteWorkflowVisibilityRecordRequest{
-		NamespaceId: s.namespaceID.String(),
-		Execution: &commonpb.WorkflowExecution{
-			WorkflowId: execution.WorkflowId,
-			RunId:      runID,
-		},
-	}).Return(&historyservice.DeleteWorkflowVisibilityRecordResponse{}, nil)
-	s.mockExecutionMgr.EXPECT().DeleteCurrentWorkflowExecution(gomock.Any(), &persistence.DeleteCurrentWorkflowExecutionRequest{
-		ShardID:     shardID,
-		NamespaceID: s.namespaceID.String(),
-		WorkflowID:  execution.WorkflowId,
-		RunID:       runID,
-	}).Return(nil)
-	s.mockExecutionMgr.EXPECT().DeleteWorkflowExecution(gomock.Any(), &persistence.DeleteWorkflowExecutionRequest{
-		ShardID:     shardID,
-		NamespaceID: s.namespaceID.String(),
-		WorkflowID:  execution.WorkflowId,
-		RunID:       runID,
-	}).Return(nil)
-	s.mockExecutionMgr.EXPECT().DeleteHistoryBranch(gomock.Any(), gomock.Any()).Times(len(mutableState.ExecutionInfo.VersionHistories.Histories))
-
-	_, err = s.handler.DeleteWorkflowExecution(context.Background(), request)
-	s.NoError(err)
-}
-
-func (s *adminHandlerSuite) TestDeleteWorkflowExecution_LoadMutableStateFailed() {
-	execution := commonpb.WorkflowExecution{
-		WorkflowId: "workflowID",
-		RunId:      uuid.New(),
-	}
-
-	request := &adminservice.DeleteWorkflowExecutionRequest{
-		Namespace: s.namespace.String(),
-		Execution: &execution,
-	}
-
-	s.mockNamespaceCache.EXPECT().GetNamespaceID(s.namespace).Return(s.namespaceID, nil).AnyTimes()
-	s.mockVisibilityMgr.EXPECT().HasStoreName(cassandra.CassandraPersistenceName).Return(false)
-
-	s.mockExecutionMgr.EXPECT().GetWorkflowExecution(gomock.Any(), gomock.Any()).Return(nil, errors.New("some random error"))
-	s.mockHistoryClient.EXPECT().DeleteWorkflowVisibilityRecord(gomock.Any(), gomock.Any()).Return(&historyservice.DeleteWorkflowVisibilityRecordResponse{}, nil)
-	s.mockExecutionMgr.EXPECT().DeleteCurrentWorkflowExecution(gomock.Any(), gomock.Any()).Return(nil)
-	s.mockExecutionMgr.EXPECT().DeleteWorkflowExecution(gomock.Any(), gomock.Any()).Return(nil)
-
-	_, err := s.handler.DeleteWorkflowExecution(context.Background(), request)
-	s.NoError(err)
-}
-
-func (s *adminHandlerSuite) TestDeleteWorkflowExecution_CassandraVisibilityBackend() {
-	execution := commonpb.WorkflowExecution{
-		WorkflowId: "workflowID",
-		RunId:      uuid.New(),
-	}
-
-	request := &adminservice.DeleteWorkflowExecutionRequest{
-		Namespace: s.namespace.String(),
-		Execution: &execution,
-	}
-
-	s.mockNamespaceCache.EXPECT().GetNamespaceID(s.namespace).Return(s.namespaceID, nil).AnyTimes()
-	s.mockVisibilityMgr.EXPECT().HasStoreName(cassandra.CassandraPersistenceName).Return(true).AnyTimes()
-
-	// test delete open records
-	branchToken := []byte("branchToken")
-	version := int64(100)
-	mutableState := &persistencespb.WorkflowMutableState{
-		ExecutionState: &persistencespb.WorkflowExecutionState{
-			CreateRequestId: uuid.New(),
-			RunId:           execution.RunId,
-			State:           enums.WORKFLOW_EXECUTION_STATE_RUNNING,
-			Status:          enumspb.WORKFLOW_EXECUTION_STATUS_RUNNING,
-		},
-		NextEventId: 12,
-		ExecutionInfo: &persistencespb.WorkflowExecutionInfo{
-			CompletionEventBatchId: 10,
-			StartTime:              timestamp.TimePtr(time.Now()),
-			VersionHistories: &historyspb.VersionHistories{
-				CurrentVersionHistoryIndex: 0,
-				Histories: []*historyspb.VersionHistory{
-					{
-						BranchToken: branchToken,
-						Items: []*historyspb.VersionHistoryItem{
-							{EventId: 11, Version: version},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	s.mockExecutionMgr.EXPECT().GetWorkflowExecution(gomock.Any(), gomock.Any()).Return(&persistence.GetWorkflowExecutionResponse{State: mutableState}, nil)
-	s.mockHistoryClient.EXPECT().DeleteWorkflowVisibilityRecord(gomock.Any(), &historyservice.DeleteWorkflowVisibilityRecordRequest{
-		NamespaceId:       s.namespaceID.String(),
-		Execution:         &execution,
-		WorkflowStartTime: mutableState.ExecutionInfo.StartTime,
-	}).Return(&historyservice.DeleteWorkflowVisibilityRecordResponse{}, nil)
-	s.mockExecutionMgr.EXPECT().DeleteCurrentWorkflowExecution(gomock.Any(), gomock.Any()).Return(nil)
-	s.mockExecutionMgr.EXPECT().DeleteWorkflowExecution(gomock.Any(), gomock.Any()).Return(nil)
-	s.mockExecutionMgr.EXPECT().DeleteHistoryBranch(gomock.Any(), gomock.Any()).Times(len(mutableState.ExecutionInfo.VersionHistories.Histories))
-
-	_, err := s.handler.DeleteWorkflowExecution(context.Background(), request)
-	s.NoError(err)
-
-	// test delete close records
-	mutableState.ExecutionState.State = enums.WORKFLOW_EXECUTION_STATE_COMPLETED
-	mutableState.ExecutionState.Status = enumspb.WORKFLOW_EXECUTION_STATUS_COMPLETED
-
-	shardID := common.WorkflowIDToHistoryShard(
-		s.namespaceID.String(),
-		execution.GetWorkflowId(),
-		s.handler.numberOfHistoryShards,
-	)
-	closeTime := time.Now()
-	s.mockExecutionMgr.EXPECT().GetWorkflowExecution(gomock.Any(), gomock.Any()).Return(&persistence.GetWorkflowExecutionResponse{State: mutableState}, nil)
-	s.mockExecutionMgr.EXPECT().ReadHistoryBranch(gomock.Any(), &persistence.ReadHistoryBranchRequest{
-		ShardID:     shardID,
-		BranchToken: branchToken,
-		MinEventID:  mutableState.ExecutionInfo.CompletionEventBatchId,
-		MaxEventID:  mutableState.NextEventId,
-		PageSize:    1,
-	}).Return(&persistence.ReadHistoryBranchResponse{
-		HistoryEvents: []*historypb.HistoryEvent{
-			{
-				EventId:   10,
-				EventType: enumspb.EVENT_TYPE_WORKFLOW_TASK_COMPLETED,
-				Version:   version,
-				EventTime: timestamp.TimePtr(closeTime.Add(-time.Millisecond)),
-			},
-			{
-				EventId:   11,
-				EventType: enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_COMPLETED,
-				Version:   version,
-				EventTime: timestamp.TimePtr(closeTime),
-			},
-		},
-	}, nil)
-	s.mockHistoryClient.EXPECT().DeleteWorkflowVisibilityRecord(gomock.Any(), &historyservice.DeleteWorkflowVisibilityRecordRequest{
-		NamespaceId:       s.namespaceID.String(),
-		Execution:         &execution,
-		WorkflowCloseTime: timestamp.TimePtr(closeTime),
-	}).Return(&historyservice.DeleteWorkflowVisibilityRecordResponse{}, nil)
-	s.mockExecutionMgr.EXPECT().DeleteCurrentWorkflowExecution(gomock.Any(), gomock.Any()).Return(nil)
-	s.mockExecutionMgr.EXPECT().DeleteWorkflowExecution(gomock.Any(), gomock.Any()).Return(nil)
-	s.mockExecutionMgr.EXPECT().DeleteHistoryBranch(gomock.Any(), gomock.Any()).Times(len(mutableState.ExecutionInfo.VersionHistories.Histories))
-
-	_, err = s.handler.DeleteWorkflowExecution(context.Background(), request)
-	s.NoError(err)
 }
 
 func (s *adminHandlerSuite) TestStreamWorkflowReplicationMessages_ClientToServerBroken() {
